Pull the factory class to package level
